+++ src/pages/Comment.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 
 import { getPostByPostId } from '@/api/fetchPosts';
-import CommentSection from '@/components/comment/commentSection';
+import CommentSection from '@/components/comment/CommentSection';
 import BackHeader from '@/components/layout/header/BackHeader';
 import Post from '@/components/post/Post';
 import { PostModel } from '@/types/post';
